requirements.txt
attrdict==2.0.0
PyYAML==3.12
pytest==3.5.0
flake8==3.5.0
setuptools>=38.5.1
